samba/smb.conf: Remove socket options
The "socket options" was just a leftover from old versions, and the default
settings on newer versions are already tuned.

The samba Wiki has this to say about it:
Modern UNIX operating systems are tuned for high network performance by default. For example, Linux has an auto-tuning mechanism for buffer sizes. When you set the socket options parameter in the smb.conf file, you are overriding these settings. In most cases, setting this parameter decreases the performance.