python3: save byte-code to their legacy locations and names
This commit patches the importlib to safe by default the byte-code
to the legacy location and name.

That means we are not longer create __pycache__ folder but instead
we are going to save the byte-code next to .py file as .pyc.

That change required in order to continue providing sourceless
installation of the plugins and significantly reduce the image size.

Additionally we restore the default optimization level to 1 just
like we used to all those years. The optimization level 1 will
remove assert and __debug__ but the extention will remain .pyc
instead of .pyo